moved the Scheduler trait to a concurrent package to avoid clashes with Monix's Scheduler
