--- jvm/src/main/scala/org/atnos/eff/Schedulers.scala
@@ -1,3 +1,0 @@
-package org.atnos.eff
-
-trait Schedulers
