+++ jvm/src/main/scala/org/atnos/eff/concurrent/Schedulers.scala
@@ -1,0 +1,3 @@
+package org.atnos.eff.concurrent
+
+trait Schedulers
